결제 관리자 페이지 생성

+++ navi-project-boot/src/main/java/com/navi/user/dto/admin/AdminDashboardDTO.java
@@ -2,6 +2,7 @@ package com.navi.user.dto.admin;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -77,7 +78,7 @@ public class AdminDashboardDTO {
     @AllArgsConstructor
     public static class Payments {
         private long count;
-        private long amount;
+        private BigDecimal amount;
         private double changedPct;
     }
